refactorisation et debut de l'écriture des tests
