+++ com.democratia.test/ViewModels/internaute/CreerGroupe/PremierePageViewModelTest.cs
@@ -1,0 +1,6 @@
+﻿namespace com.democratia.test.ViewModels.internaute.CreerGroupe
+{
+    public class PremierePageViewModelTest
+    {
+    }
+}
